PAPP-4096 various changes to support MAS 2.2
Also addresses PAPP-4105, PAPP-4095, and PAPP-4097.

The header name for the API key seems to have changed from "token" to
"X-API-TOKEN" so I changed it.

The test_connectivity api required admin so I changed it to listing
tasks.

The risk score we were using wasn't the main one used in the MAS UI so
I changed it.

The default source of "phantom" wasn't working so I changed it to
"www", which does work.